Universal cache, simplified some code, updated TODO, NSWDB rev support
Cache is handled by the XCI PackageID. Luckily, PackageIDs are unique and quick to get. This makes the cache universal since it only stores the PackageID and TitleID. Thanks LucaFraga for hacbuild code!
All of these changes may make the program slightly slower because the file sizes are not cached anymore (as well as the overhead for PackageID).